Destiny's Oasis V1.05
Added enemy attacks, and damage. Now you can kill the enemies by clicking on them, there is also now a score to show the time youve survived as well as the amount youve killed

+++ Destiny's Oasis/Assets/Scripts/Player/playerHealth.cs
@@ -16,7 +16,7 @@ public class playerHealth : MonoBehaviour
         }
     }
 
-    public void TakeDamage(float Damage)
+    public void TakeDamage(int Damage)
     {
         healthAmount -= Damage;
         healthBar.fillAmount = healthAmount / 100;
